Move all classes under DocumentFormat.OpenXml.Packaging
