+++ DocumentFormat.OpenXml/Packaging/DataPart.cs
@@ -2,7 +2,6 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Text;
 using System.Diagnostics;
 using System.IO;
 using System.IO.Packaging;
@@ -321,319 +320,6 @@ namespace DocumentFormat.OpenXml.Packaging
         }
 
         #endregion
-    }
-
-    /// <summary>
-    /// Represents a media (Audio, Video) data part in the document.
-    /// </summary>
-    public class MediaDataPart : DataPart
-    {
-        private const string DefaultTargetPart = "media";
-        private const string DefaultTargetName = "mediadata";
-        private const string DefaultTargetExt = ".bin";
-
-        /// <summary>
-        /// Initializes a new instance of the MediaDataPart class. 
-        /// </summary>
-        internal MediaDataPart()
-            : base()
-        {
-        }
-
-        /// <summary>
-        /// The internal path to be used for the part name.
-        /// </summary>
-        internal override string TargetPath
-        {
-            get { return DefaultTargetPart; }
-        }
-
-        /// <summary>
-        /// The file base name to be used for the part name in the package
-        /// </summary>
-        internal override string TargetName
-        {
-            get { return DefaultTargetName; }
-        }
-
-        /// <summary>
-        /// The file extension to be used for the part in the package.
-        /// </summary>
-        internal override string TargetFileExtension
-        {
-            get { return DefaultTargetExt; }
-        }
-    }
-
-    /// <summary>
-    /// Defines part media types.
-    /// </summary>
-    public enum MediaDataPartType
-    {
-        #region Audio Types
-
-        /// <summary>
-        /// Audio Interchange File Format (.aiff)
-        /// </summary>
-        Aiff,
-        ///// <summary>
-        ///// AU Audio File (.au)
-        ///// </summary>
-        //Basic, 
-        /// <summary>
-        /// MIDI Audio (.mid)
-        /// </summary>
-        Midi,
-        /// <summary>
-        /// MP3 (.mp3)
-        /// </summary>
-        Mp3,
-        /// <summary>
-        /// MP3 Playlist File (.m3u)
-        /// </summary>
-        MpegUrl,
-        /// <summary>
-        /// WAV audio (.wav)
-        /// </summary>
-        Wav,
-        ///// <summary>
-        ///// Windows Media Audio Redirector (.wax)
-        ///// </summary>
-        //Wax,
-        /// <summary>
-        /// Windows Media Audio File (.wma)
-        /// </summary>
-        Wma,
-        /// <summary>
-        /// Mpeg audio (.mpeg)
-        /// </summary>
-        MpegAudio,
-        /// <summary>
-        /// Ogg Vorbis (.ogg)
-        /// </summary>
-        OggAudio,
-
-        #endregion
-
-        #region Video Types
-
-        ///// <summary>
-        ///// Advanced Streaming Format (.asf)
-        ///// </summary>
-        //Asf,
-        /// <summary>
-        /// Advanced Stream Redirector File (.asx)
-        /// </summary>
-        Asx,
-        /// <summary>
-        /// Audio Video Interleave File (.avi)
-        /// </summary>
-        Avi,
-        ///// <summary>
-        ///// Indeo Video File (.ivf)
-        ///// </summary>
-        //Ivf,
-        /// <summary>
-        /// MPEG 1 System Stream (.mpg)
-        /// </summary>
-        Mpg,
-        /// <summary>
-        /// MPEG 1 System Stream (.mpeg)
-        /// </summary>
-        MpegVideo,
-        ///// <summary>
-        ///// Windows Media A/V File (.wm)
-        ///// </summary>
-        //WM,
-        /// <summary>
-        /// Windows Media File (.wmv)
-        /// </summary>
-        Wmv,
-        /// <summary>
-        /// Windows Media Player A/V Shortcut (.wmx)
-        /// </summary>
-        Wmx,
-        /// <summary>
-        /// Windows Media Redirector (.wvx)
-        /// </summary>
-        Wvx,
-        /// <summary>
-        /// QuickTime video (.mov)
-        /// </summary>
-        Quicktime,
-        /// <summary>
-        ///  Ogg Stream (.ogg)
-        /// </summary>
-        OggVideo,
-        /// <summary>
-        /// VC-1 Stream (.wmv)
-        /// </summary>
-        VC1
-
-        #endregion
-    }
-
-    internal static class MediaDataPartTypeInfo
-    {
-        internal static string GetContentType(this MediaDataPartType mediaDataPartType)
-        {
-            switch (mediaDataPartType)
-            {
-                case MediaDataPartType.Aiff:
-                    return "audio/aiff";
-
-                //case MediaDataPartType.Basic:
-                //    return "audio/basic";
-
-                case MediaDataPartType.Midi:
-                    return "audio/midi";
-
-                case MediaDataPartType.Mp3:
-                    return "audio/mp3";
-
-                case MediaDataPartType.MpegUrl:
-                    return "audio/mpegurl";
-
-                case MediaDataPartType.Wav:
-                    return "audio/wav";
-
-                //case MediaDataPartType.Wax:
-                //    return "audio/x-ms-wax";
-
-                case MediaDataPartType.Wma:
-                    return "audio/x-ms-wma";
-
-                case MediaDataPartType.MpegAudio:
-                    return "audio/mpeg";
-
-                case MediaDataPartType.OggAudio:
-                    return "audio/ogg";
-
-                // *** Video types.
-
-                //case MediaDataPartType.Asf:
-                //    return "video/x-ms-asf";
-
-                case MediaDataPartType.Asx:
-                    return "video/x-ms-asf-plugin";
-
-                case MediaDataPartType.Avi:
-                    return "video/avi";
-
-                //case MediaDataPartType.Ivf:
-                //    return "video/x-ivf";
-
-                case MediaDataPartType.Mpg:
-                    return "video/mpg";
-
-                case MediaDataPartType.MpegVideo:
-                    return "video/mpeg";
-
-                //case MediaDataPartType.WM:
-                //    return "video/x-ms-wm";
-
-                case MediaDataPartType.Wmv:
-                    return "video/x-ms-wmv";
-
-                case MediaDataPartType.Wmx:
-                    return "video/x-ms-wmx";
-
-                case MediaDataPartType.Wvx:
-                    return "video/x-ms-wvx";
-
-                case MediaDataPartType.Quicktime:
-                    return "video/quicktime";
-
-                case MediaDataPartType.OggVideo:
-                    return "video/ogg";
-
-                case MediaDataPartType.VC1:
-                    return "video/vc1";
-
-                default:
-                    throw new ArgumentOutOfRangeException(nameof(mediaDataPartType));
-            }
-        }
-
-        internal static string GetTargetExtension(this MediaDataPartType mediaDataPartType)
-        {
-            switch (mediaDataPartType)
-            {
-                case MediaDataPartType.Aiff:
-                    return ".aiff";
-
-                //case MediaDataPartType.Basic:
-                //    return ".au";
-
-                case MediaDataPartType.Midi:
-                    return ".midi";
-
-                case MediaDataPartType.Mp3:
-                    return ".mp3";
-
-                case MediaDataPartType.MpegUrl:
-                    return ".m3u";
-
-                case MediaDataPartType.Wav:
-                    return ".wav";
-
-                //case MediaDataPartType.Wax:
-                //    return ".wax";
-
-                case MediaDataPartType.Wma:
-                    return ".wma";
-
-                case MediaDataPartType.MpegAudio:
-                    return ".mpeg";
-
-                case MediaDataPartType.OggAudio:
-                    return ".ogg";
-
-                // Video types.
-
-                //case MediaDataPartType.Asf:
-                //    return ".asf";
-
-                case MediaDataPartType.Asx:
-                    return ".asx";
-
-                case MediaDataPartType.Avi:
-                    return ".avi";
-
-                //case MediaDataPartType.Ivf:
-                //    return ".ivf";
-
-                case MediaDataPartType.Mpg:
-                    return ".mpg";
-
-                case MediaDataPartType.MpegVideo:
-                    return ".mpeg";
-
-                //case MediaDataPartType.WM:
-                //    return ".wm";
-
-                case MediaDataPartType.Wmv:
-                    return ".wmv";
-
-                case MediaDataPartType.Wmx:
-                    return ".wmx";
-
-                case MediaDataPartType.Wvx:
-                    return ".wvx";
-
-                case MediaDataPartType.Quicktime:
-                    return ".mov";
-
-                case MediaDataPartType.OggVideo:
-                    return ".ogg";
-
-                case MediaDataPartType.VC1:
-                    return ".wmv";
-                
-                default:
-                    return ".media";
-            }
-        }
     }
 
 }
