+++ DocumentFormat.OpenXml/Packaging/ExtendedPart.cs
@@ -1,8 +1,7 @@
 ﻿// Copyright (c) Microsoft Open Technologies, Inc.  All rights reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
+
 using System;
 using System.Collections.Generic;
-using System.Text;
-using System.IO.Packaging;
 
 namespace DocumentFormat.OpenXml.Packaging
 {
