+++ DocumentFormat.OpenXml/Packaging/IFixedTypePart.cs
