+++ DocumentFormat.OpenXml/Packaging/OpenXmlBasePart.cs
@@ -1,4 +1,5 @@
 ﻿// Copyright (c) Microsoft Open Technologies, Inc.  All rights reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
+
 using System;
 using System.IO;
 using System.Collections.Generic;
