+++ DocumentFormat.OpenXml/Packaging/OpenXmlObsolete.cs
