+++ DocumentFormat.OpenXml/Packaging/OpenXmlPackage.cs
@@ -1,151 +1,18 @@
 ﻿// Copyright (c) Microsoft Open Technologies, Inc.  All rights reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
-// Copyright 2014 Thomas Barnekow (cloning, Flat OPC (with Eric White))
+
 using System;
-using System.Diagnostics;
-using System.Collections;
 using System.Collections.Generic;
-using System.Collections.ObjectModel;
-using System.Collections.Specialized;
-using System.Text;
+using System.Diagnostics;
+using System.Globalization;
 using System.IO;
 using System.IO.Packaging;
-using System.Globalization;
-using DocumentFormat.OpenXml;
 using System.Linq;
+using System.Text;
 using System.Xml;
 using System.Xml.Linq;
 
-#if FEATURE_SERIALIZATION
-using System.Runtime.Serialization;
-#endif
-
 namespace DocumentFormat.OpenXml.Packaging
 {
-    internal struct RelationshipProperty
-    {
-        internal string Id;
-        internal string RelationshipType;
-        internal TargetMode TargetMode;
-        internal Uri TargetUri;
-    };
-
-    /// <summary>
-    /// Defines the base class for PackageRelationshipPropertyCollection and PackagePartRelationshipPropertyCollection objects.
-    /// </summary>
-    abstract internal class RelationshipCollection : List<RelationshipProperty>
-    {
-        protected PackageRelationshipCollection BasePackageRelationshipCollection { get; set; }
-
-        internal bool StrictTranslation { get; set; }
-
-        /// <summary>
-        /// This method fills the collection with PackageRels from the PackageRelationshipCollection that is given in the sub class.
-        /// </summary>
-        protected void Build()
-        {
-            foreach (PackageRelationship relationship in this.BasePackageRelationshipCollection)
-            {
-                bool found;
-                string transitionalNamespace;
-                RelationshipProperty relationshipProperty;
-
-                relationshipProperty.TargetUri = relationship.TargetUri;
-                relationshipProperty.TargetMode = relationship.TargetMode;
-                relationshipProperty.Id = relationship.Id;
-                relationshipProperty.RelationshipType = relationship.RelationshipType;
-
-                // If packageRel.RelationshipType is something for Strict, it tries to get the equivalent in Transitional.
-                found = NamespaceIdMap.TryGetTransitionalRelationship(relationshipProperty.RelationshipType, out transitionalNamespace);
-                if (found)
-                {
-                    relationshipProperty.RelationshipType = transitionalNamespace;
-                    this.StrictTranslation = true;
-                }
-
-                this.Add(relationshipProperty);
-            }
-        }
-
-        internal void UpdateRelationshipTypesInPackage()
-        {
-            // Update the relationshipTypes when editable.
-            if (this.GetPackage().FileOpenAccess != FileAccess.Read)
-            {
-                for (int index = 0; index < this.Count; index++)
-                {
-                    RelationshipProperty relationshipProperty = this[index];
-                    this.ReplaceRelationship(relationshipProperty.TargetUri, relationshipProperty.TargetMode, relationshipProperty.RelationshipType, relationshipProperty.Id);
-                }
-            }
-        }
-
-        abstract internal void ReplaceRelationship(Uri targetUri, TargetMode targetMode, string strRelationshipType, string strId);
-
-        abstract internal Package GetPackage();
-    }
-
-    /// <summary>
-    /// Represents a collection of relationships that are obtained from the package.
-    /// </summary>
-    internal class PackageRelationshipPropertyCollection : RelationshipCollection
-    {
-        public Package BasePackage { get; set; }
-
-        public PackageRelationshipPropertyCollection(Package package)
-        {
-            this.BasePackage = package;
-            if (this.BasePackage == null)
-            {
-                throw new ArgumentNullException(nameof(BasePackage));
-            }
-
-            this.BasePackageRelationshipCollection = this.BasePackage.GetRelationships();
-            this.Build();
-        }
-
-        internal override void ReplaceRelationship(Uri targetUri, TargetMode targetMode, string strRelationshipType, string strId)
-        {
-            this.BasePackage.DeleteRelationship(strId);
-            this.BasePackage.CreateRelationship(targetUri, targetMode, strRelationshipType, strId);
-        }
-
-        internal override Package GetPackage()
-        {
-            return this.BasePackage;
-        }
-    }
-
-    /// <summary>
-    /// Represents a collection of relationships that are obtained from the package part.
-    /// </summary>
-    internal class PackagePartRelationshipPropertyCollection : RelationshipCollection
-    {
-        public PackagePart BasePackagePart { get; set; }
-
-        public PackagePartRelationshipPropertyCollection(PackagePart packagePart)
-        {
-            this.BasePackagePart = packagePart;
-            if (this.BasePackagePart == null)
-            {
-                throw new ArgumentNullException(nameof(BasePackagePart));
-            }
-
-            this.BasePackageRelationshipCollection = this.BasePackagePart.GetRelationships();
-            this.Build();
-        }
-
-        internal override void ReplaceRelationship(Uri targetUri, TargetMode targetMode, string strRelationshipType, string strId)
-        {
-            this.BasePackagePart.DeleteRelationship(strId);
-            this.BasePackagePart.CreateRelationship(targetUri, targetMode, strRelationshipType, strId);
-        }
-
-        internal override Package GetPackage()
-        {
-            return this.BasePackagePart.Package;
-        }
-    }
-
     /// <summary>
     /// Represents a base class for strong typed Open XML document classes.
     /// </summary>
@@ -1367,147 +1234,7 @@ namespace DocumentFormat.OpenXml.Packaging
             return dataPart;
         }
 
-        #endregion
-
-        internal class PartUriHelper
-        {
-            private Dictionary<string, int> _sequenceNumbers = new Dictionary<string, int>(20);
-            private Dictionary<string, int> _reservedUri = new Dictionary<string, int>();
-
-            //List of contentTypes that need to have a 1 appended to the name
-            //for the first item in the package. Section numbers in comments
-            //refer to the ISO/IEC 29500 standard.
-            private static readonly HashSet<string> _numberedContentTypes = new HashSet<string>(StringComparer.OrdinalIgnoreCase)
-            {
-                //11.3 WordprocessingML Parts
-                "application/vnd.openxmlformats-officedocument.wordprocessingml.footer+xml",
-                "application/vnd.openxmlformats-officedocument.wordprocessingml.header+xml",
-                //12.3 SpreadsheetML Parts
-                "application/vnd.openxmlformats-officedocument.spreadsheetml.chartsheet+xml",
-                "application/vnd.openxmlformats-officedocument.spreadsheetml.comments+xml",
-                "application/vnd.openxmlformats-officedocument.spreadsheetml.dialogsheet+xml",
-                "application/vnd.openxmlformats-officedocument.drawing+xml",
-                "application/vnd.openxmlformats-officedocument.spreadsheetml.externalLink+xml",
-                "application/vnd.openxmlformats-officedocument.spreadsheetml.dialogsheet+xml",
-                "application/vnd.openxmlformats-officedocument.drawing+xml",
-                "application/vnd.openxmlformats-officedocument.spreadsheetml.externalLink+xml",
-                "application/vnd.openxmlformats-officedocument.spreadsheetml.sheetMetadata+xml",
-                "application/vnd.openxmlformats-officedocument.spreadsheetml.pivotCacheDefinition+xml",
-                "application/vnd.openxmlformats-officedocument.spreadsheetml.pivotCacheRecords+xml",
-                "application/vnd.openxmlformats-officedocument.spreadsheetml.queryTable+xml",
-                "application/vnd.openxmlformats-officedocument.spreadsheetml.revisionLog+xml",
-                "application/vnd.openxmlformats-officedocument.spreadsheetml.tableSingleCells+xml",
-                "application/vnd.openxmlformats-officedocument.spreadsheetml.table+xml",
-                "application/vnd.openxmlformats-officedocument.spreadsheetml.worksheet+xml",
-                //13.3 PresentationML Parts
-                "application/vnd.openxmlformats-officedocument.presentationml.comments+xml",
-                "application/vnd.openxmlformats-officedocument.presentationml.handoutMaster+xml",
-                "application/vnd.openxmlformats-officedocument.presentationml.notesMaster+xml",
-                "application/vnd.openxmlformats-officedocument.presentationml.notesSlide+xml",
-                "application/vnd.openxmlformats-officedocument.presentationml.slide+xml",
-                "application/vnd.openxmlformats-officedocument.presentationml.slideLayout+xml",
-                "application/vnd.openxmlformats-officedocument.presentationml.slideMaster+xml",
-                "application/vnd.openxmlformats-officedocument.presentationml.slideUpdateInfo+xml",
-                "application/vnd.openxmlformats-officedocument.presentationml.tags+xml",
-                //14.2 DrawingML Parts
-                "application/vnd.openxmlformats-officedocument.drawingml.chart+xml",
-                "application/vnd.openxmlformats-officedocument.drawingml.chartshapes+xml",
-                "application/vnd.openxmlformats-officedocument.drawingml.diagramColors+xml",
-                "application/vnd.openxmlformats-officedocument.drawingml.diagramData+xml",
-                "application/vnd.openxmlformats-officedocument.drawingml.diagramLayout+xml",
-                "application/vnd.openxmlformats-officedocument.drawingml.diagramStyle+xml",
-                "application/vnd.openxmlformats-officedocument.theme+xml",
-                "application/vnd.openxmlformats-officedocument.themeOverride+xml",
-                //15.2 Shared Parts
-                "application/vnd.openxmlformats-officedocument.customXmlProperties+xml",
-                "application/vnd.openxmlformats-officedocument.spreadsheetml.printerSettings",
-                "application/vnd.openxmlformats-officedocument.wordprocessingml.printerSettings",
-                "application/vnd.openxmlformats-officedocument.presentationml.printerSettings"
-            };
-
-            public PartUriHelper()
-            {
-            }
-
-            private bool IsReservedUri(Uri uri)
-            {
-                string uriString = uri.OriginalString.ToUpperInvariant();
-                return this._reservedUri.ContainsKey(uriString);
-            }
-
-            internal void AddToReserveUri(Uri partUri)
-            {
-                string uriString = partUri.OriginalString.ToUpperInvariant();
-                this._reservedUri.Add(uriString, 0);
-            }
-
-            internal void ReserveUri(string contentType, Uri partUri)
-            {
-                GetNextSequenceNumber(contentType);
-                this.AddToReserveUri(PackUriHelper.GetNormalizedPartUri(partUri));
-            }
-
-            internal Uri GetUniquePartUri(string contentType, Uri parentUri, string targetPath, string targetName, string targetExt)
-            {
-                Uri partUri;
-
-                do
-                {
-                    string sequenceNumber = this.GetNextSequenceNumber(contentType);
-                    string path = Path.Combine(targetPath, targetName + sequenceNumber + targetExt);
-
-                    Uri uri = new Uri(path, UriHelper.RelativeOrAbsolute);
-                    partUri = PackUriHelper.ResolvePartUri(parentUri, uri);
-                    // partUri = PackUriHelper.GetNormalizedPartUri(PackUriHelper.CreatePartUri(uri));
-                } while (this.IsReservedUri(partUri));
-
-                this.AddToReserveUri(partUri);
-
-                // do not need to add to the _existedNames
-                return partUri;
-            }
-
-            internal Uri GetUniquePartUri(string contentType, Uri parentUri, Uri targetUri)
-            {
-                Uri partUri;
-
-                partUri = PackUriHelper.ResolvePartUri(parentUri, targetUri);
-
-                if (this.IsReservedUri(partUri))
-                {
-                    // already have one, create new
-                    string targetPath = ".";
-                    string targetName = Path.GetFileNameWithoutExtension(targetUri.OriginalString);
-                    string targetExt = Path.GetExtension(targetUri.OriginalString);
-
-                    partUri = GetUniquePartUri(contentType, partUri, targetPath, targetName, targetExt);
-                }
-                else
-                {
-                    // not used, can use it.
-                    this.AddToReserveUri(partUri);
-                }
-
-                return partUri;
-            }
-
-            private string GetNextSequenceNumber(string contentType)
-            {
-                if (this._sequenceNumbers.ContainsKey(contentType))
-                {
-                    this._sequenceNumbers[contentType] += 1;
-                    // use the default read-only NumberFormatInfo that is culture-independent (invariant). 
-                    return this._sequenceNumbers[contentType].ToString(NumberFormatInfo.InvariantInfo);
-                }
-                else
-                {
-                    this._sequenceNumbers.Add(contentType, 1);
-
-                    //Certain contentTypes need to be numbered starting with 1.
-                    return _numberedContentTypes.Contains(contentType) ? "1" : "";
-                }
-            }
-        }
+#endregion
 
         #region saving and cloning
 
@@ -2053,368 +1780,5 @@ namespace DocumentFormat.OpenXml.Packaging
         }
 
         #endregion Flat OPC
-    }
-
-    /// <summary>
-    /// Specifies event handlers that will handle OpenXmlPackage validation events and the OpenXmlPackageValidationEventArgs. 
-    /// </summary>
-    // Building on Travis CI failed, saying that ObsoleteAttributeMessages does not contain a definition
-    // for 'ObsoleteV1ValidationFunctionality'. Thus, we've replaced the member with its value.
-    [Obsolete("This functionality is obsolete and will be removed from future version release. Please see OpenXmlValidator class for supported validation functionality.", false)]
-    public class OpenXmlPackageValidationSettings
-    {
-        private EventHandler<OpenXmlPackageValidationEventArgs> valEventHandler;
-
-        /// <summary>
-        /// Gets the event handler.
-        /// </summary>
-        /// <returns></returns>
-        internal EventHandler<OpenXmlPackageValidationEventArgs> GetEventHandler()
-        {
-            return this.valEventHandler;
-        }
-
-        /// <summary>
-        /// Represents the callback method that will handle OpenXmlPackage validation events and the OpenXmlPackageValidationEventArgs.
-        /// </summary>
-        public event EventHandler<OpenXmlPackageValidationEventArgs> EventHandler
-        {
-            add
-            {
-                this.valEventHandler = (EventHandler<OpenXmlPackageValidationEventArgs>)Delegate.Combine(this.valEventHandler, value);
-            }
-
-            remove
-            {
-                this.valEventHandler = (EventHandler<OpenXmlPackageValidationEventArgs>)Delegate.Remove(this.valEventHandler, value);
-            }
-        }
-
-        /// <summary>
-        /// Gets or sets the file format version that the validation is targeting.
-        /// </summary>
-        internal FileFormatVersions FileFormat
-        {
-            get;
-            set;
-        }
-    }
-
-    /// <summary>
-    /// Represents the Open XML package validation events.
-    /// </summary>
-    [SerializableAttribute]
-    [Obsolete(ObsoleteAttributeMessages.ObsoleteV1ValidationFunctionality, false)]
-    public sealed class OpenXmlPackageValidationEventArgs : EventArgs
-    {
-        private string _message;
-        private string _partClassName;
-
-        [NonSerializedAttribute]
-        private OpenXmlPart _childPart;
-
-        [NonSerializedAttribute]
-        private OpenXmlPart _parentPart;
-
-        internal OpenXmlPackageValidationEventArgs()
-        {
-        }
-
-        /// <summary>
-        /// Gets the message string of the event.
-        /// </summary>
-        public string Message
-        {
-            get
-            {
-                if (this._message == null && this.MessageId != null)
-                {
-                    return ExceptionMessages.ResourceManager.GetString(this.MessageId);
-                }
-                else
-                {
-                    return this._message;
-                }
-            }
-
-            set
-            {
-                this._message = value;
-            }
-        }
-
-        /// <summary>
-        /// Gets the class name of the part.
-        /// </summary>
-        public string PartClassName
-        {
-            get { return _partClassName; }
-            internal set { _partClassName = value; }
-        }
-
-        /// <summary>
-        /// Gets the part that caused the event.
-        /// </summary>
-        public OpenXmlPart SubPart
-        {
-            get { return _childPart; }
-            internal set { _childPart = value; }
-        }
-
-        /// <summary>
-        /// Gets the part in which to process the validation.
-        /// </summary>
-        public OpenXmlPart Part
-        {
-            get { return _parentPart; }
-            internal set { _parentPart = value; }
-        }
-
-        internal string MessageId
-        {
-            get;
-            set;
-        }
-
-        /// <summary>
-        /// The DataPartReferenceRelationship that caused the event.
-        /// </summary>
-        internal DataPartReferenceRelationship DataPartReferenceRelationship
-        {
-            get;
-            set;
-        }
-    }
-
-    /// <summary>
-    /// Represents an Open XML package exception class for errors.
-    /// </summary>
-    [SerializableAttribute]
-    public sealed class OpenXmlPackageException : Exception
-    {
-        /// <summary>
-        /// Initializes a new instance of the OpenXmlPackageException class. 
-        /// </summary>
-        public OpenXmlPackageException()
-            : base()
-        {
-        }
-
-        /// <summary>
-        /// Initializes a new instance of the OpenXmlPackageException class using the supplied error message. 
-        /// </summary>
-        /// <param name="message">The message that describes the error. </param>
-        public OpenXmlPackageException(string message)
-            : base(message)
-        {
-        }
-
-#if FEATURE_SERIALIZATION
-        /// <summary>
-        /// Initializes a new instance of the OpenXmlPackageException class using the supplied serialized data. 
-        /// </summary>
-        /// <param name="info">The serialized object data about the exception being thrown.</param>
-        /// <param name="context">The contextual information about the source or destination.</param>
-        private OpenXmlPackageException(SerializationInfo info, StreamingContext context)
-            : base(info, context)
-        {
-        }
-#endif
-
-        /// <summary>
-        /// Initializes a new instance of the OpenXmlPackageException class using the supplied error message and a reference to the inner exception that caused the current exception. 
-        /// </summary>
-        /// <param name="message">The error message that indicates the reason for the exception.</param>
-        /// <param name="innerException">The exception that is the cause of the current exception, or a null reference (Nothing in Visual Basic) if no inner exception is specified.</param>
-        public OpenXmlPackageException(string message, Exception innerException)
-            : base(message, innerException)
-        {
-        }
-    }
-
-    /// <summary>
-    /// Represents the settings when opening a document.
-    /// </summary>
-    public class OpenSettings
-    {
-        bool? autoSave;
-        MarkupCompatibilityProcessSettings _mcSettings;
-
-        /// <summary>
-        /// Gets or sets a value that indicates whether or not to auto save document modifications.
-        /// The default value is true.
-        /// </summary>
-        public bool AutoSave
-        {
-            get
-            {
-                if (autoSave == null)
-                {
-                    return true;
-                }
-                return (bool)autoSave;
-            }
-
-            set
-            {
-                autoSave = value;
-            }
-        }
-
-        /// <summary>
-        /// Gets or sets the value of the markup compatibility processing mode.
-        /// </summary>
-        public MarkupCompatibilityProcessSettings MarkupCompatibilityProcessSettings
-        {
-            get
-            {
-                if (_mcSettings == null)
-                    _mcSettings = new MarkupCompatibilityProcessSettings(MarkupCompatibilityProcessMode.NoProcess, FileFormatVersions.Office2007);
-                return _mcSettings;
-            }
-
-            set
-            {
-                _mcSettings = value;
-            }
-        }
-
-        /// <summary>
-        /// Gets or sets a value that indicates the maximum number of allowable characters in an Open XML part. A zero (0) value indicates that there are no limits on the size of the part. A non-zero value specifies the maximum size, in characters.
-        /// </summary>
-        /// <remarks>
-        /// This property allows you to mitigate denial of service attacks where the attacker submits a package with an extremely large Open XML part. By limiting the size of the part, you can detect the attack and recover reliably.
-        /// </remarks>
-        public long MaxCharactersInPart { get; set; }
-    }
-
-    /// <summary>
-    /// Represents markup compatibility processing settings.
-    /// </summary>
-    public class MarkupCompatibilityProcessSettings
-    {
-        /// <summary>
-        /// Gets the markup compatibility process mode.
-        /// </summary>
-        public MarkupCompatibilityProcessMode ProcessMode { get; internal set; }
-
-        /// <summary>
-        /// Gets the target file format versions.
-        /// </summary>
-        public FileFormatVersions TargetFileFormatVersions { get; internal set; }
-
-        /// <summary>
-        /// Creates a MarkupCompatibilityProcessSettings object using the supplied process mode and file format versions.
-        /// </summary>
-        /// <param name="processMode">The process mode.</param>
-        /// <param name="targetFileFormatVersions">The file format versions. This parameter is ignored if the value is NoProcess.</param>
-        public MarkupCompatibilityProcessSettings(MarkupCompatibilityProcessMode processMode, FileFormatVersions targetFileFormatVersions)
-        {
-            ProcessMode = processMode;
-            TargetFileFormatVersions = targetFileFormatVersions;
-        }
-
-        private MarkupCompatibilityProcessSettings()
-        {
-            ProcessMode = MarkupCompatibilityProcessMode.NoProcess;
-            TargetFileFormatVersions = FileFormatVersions.Office2007;
-        }
-    }
-
-    /// <summary>
-    /// Specifies the mode in which to process the markup compatibility tags in the document.
-    /// </summary>
-    public enum MarkupCompatibilityProcessMode
-    {
-        /// <summary>
-        /// Do not process MarkupCompatibility tags.
-        /// </summary>
-        NoProcess = 0,
-
-        /// <summary>
-        /// Process the loaded parts.
-        /// </summary>
-        ProcessLoadedPartsOnly,
-
-        /// <summary>
-        /// Process all the parts in the package.
-        /// </summary>
-        ProcessAllParts,
-    }
-
-    /// <summary>
-    /// Traversal parts in the <see cref="OpenXmlPackage"/> by breadth-first.
-    /// </summary>
-    internal class OpenXmlPackagePartIterator : IEnumerable<OpenXmlPart>
-    {
-        private OpenXmlPackage _package;
-
-        #region Constructor
-        /// <summary>
-        /// Initializes a new instance of the OpenXmlPackagePartIterator class using the supplied OpenXmlPackage class.
-        /// </summary>
-        /// <param name="package">The OpenXmlPackage to use to enumerate parts.</param>
-        public OpenXmlPackagePartIterator(OpenXmlPackage package)
-        {
-            Debug.Assert(package != null);
-            this._package = package;
-        }
-
-        #endregion
-
-        #region IEnumerable<OpenXmlPart> Members
-        /// <summary>
-        /// Gets an enumerator for parts in the whole package.
-        /// </summary>
-        /// <returns></returns>
-        public IEnumerator<OpenXmlPart> GetEnumerator()
-        {
-            return GetPartsByBreadthFirstTraversal();
-        }
-
-        // Traverses the parts graph by breath-first
-        private IEnumerator<OpenXmlPart> GetPartsByBreadthFirstTraversal()
-        {
-            Debug.Assert(_package != null);
-            var returnedParts = new List<OpenXmlPart>();
-            Queue<OpenXmlPart> tmpQueue = new Queue<OpenXmlPart>();
-
-            // Enqueue child parts of the package.
-            foreach (var idPartPair in _package.Parts)
-            {
-                tmpQueue.Enqueue(idPartPair.OpenXmlPart);
-            }
-
-            while (tmpQueue.Count > 0)
-            {
-                var part = tmpQueue.Dequeue();
-                returnedParts.Add(part);
-
-                foreach (var subIdPartPair in part.Parts)
-                {
-                    if (!tmpQueue.Contains(subIdPartPair.OpenXmlPart)
-                        && !returnedParts.Contains(subIdPartPair.OpenXmlPart))
-                    {
-                        tmpQueue.Enqueue(subIdPartPair.OpenXmlPart);
-                    }
-                }
-            }
-            return returnedParts.GetEnumerator();
-        }
-
-        #endregion
-
-        #region IEnumerable Members
-
-        /// <summary>
-        /// Gets an enumerator for parts in the whole package.
-        /// </summary>
-        /// <returns></returns>
-        System.Collections.IEnumerator System.Collections.IEnumerable.GetEnumerator()
-        {
-            return GetEnumerator();
-        }
-
-        #endregion
     }
 }
