+++ DocumentFormat.OpenXml/Packaging/RelationshipProperty.cs
@@ -1,0 +1,15 @@
+﻿// Copyright (c) Microsoft Open Technologies, Inc.  All rights reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
+
+using System;
+using System.IO.Packaging;
+
+namespace DocumentFormat.OpenXml.Packaging
+{
+    internal struct RelationshipProperty
+    {
+        internal string Id;
+        internal string RelationshipType;
+        internal TargetMode TargetMode;
+        internal Uri TargetUri;
+    };
+}
